+++ skbuild/resources/cmake/FindPythonExtensions.cmake
@@ -53,6 +53,10 @@
 # linked modules.  They still must be forward declared and initialized, however,
 # and so are added to the forward declared modules list.
 #
+# If the associated target is of type ``MODULE_LIBRARY``, the LINK_FLAGS target
+# property is used to set symbol visibility and export only the module init function.
+# This applies to GNU and MSVC compilers.
+#
 # Options:
 #
 # ``LINKED_MODULES_VAR <LinkedModVar>``
@@ -317,6 +321,30 @@ if(NOT DEFINED PYTHON_EXTENSION_MODULE_SUFFIX)
   set(PYTHON_EXTENSION_MODULE_SUFFIX "${_item}")
 endif()
 
+function(_set_python_extension_symbol_visibility _target)
+  if(PYTHON_VERSION_MAJOR VERSION_GREATER 2)
+    set(_modinit_prefix "PyInit_")
+  else()
+    set(_modinit_prefix "init")
+  endif()
+  message("_modinit_prefix:${_modinit_prefix}")
+  if("${CMAKE_C_COMPILER_ID}" STREQUAL "MSVC")
+    set_target_properties(${_target} PROPERTIES LINK_FLAGS 
+        "/EXPORT:${_modinit_prefix}${_target}"
+    )
+  elseif("${CMAKE_C_COMPILER_ID}" STREQUAL "GNU")
+    set(_script_path
+      ${CMAKE_CURRENT_BINARY_DIR}/CMakeFiles/${_target}-version-script.map
+    )
+    file(WRITE ${_script_path} 
+               "{global: ${_modinit_prefix}${_target}; local: *; };"
+    )
+    set_property(TARGET ${_target} APPEND_STRING PROPERTY LINK_FLAGS
+        " -Wl,--version-script=${_script_path}"
+    )
+  endif()
+endfunction()
+
 function(python_extension_module _target)
   set(one_ops LINKED_MODULES_VAR FORWARD_DECL_MODULES_VAR MODULE_SUFFIX)
   cmake_parse_arguments(_args "" "${one_ops}" "" ${ARGN})
@@ -393,6 +421,10 @@ function(python_extension_module _target)
     endif()
 
     target_link_libraries_with_dynamic_lookup(${_target} ${PYTHON_LIBRARIES})
+    
+    if(_is_module_lib)
+      _set_python_extension_symbol_visibility(${_target})
+    endif()
   endif()
 endfunction()